fix: resolve major compilation errors in analytics and ehr services
- Remove duplicate Debug trait implementations in analytics reports
- Add chrono::Datelike import to EHR services for year() method access
- Fix period field type from i64 to String in analytics user_analytics
- Fix type conversions (i64 to i32) in user_analytics date handling
- Update check_type values from strings to integers in attendance modules

These changes address key compilation errors preventing the build from
succeeding with all features enabled.

+++ src/service/analytics/v1/user_analytics.rs
@@ -545,8 +545,8 @@ impl UserAnalyticsService {
                         .format("%Y-%m-%d")
                         .to_string(),
                         hour: hour as i32,
-                        day_of_week: day,
-                        week_of_year: week,
+                        day_of_week: day as i32,
+                        week_of_year: week as i32,
                         activity_score: final_score,
                         interaction_count: (final_score * 50.0) as i32,
                     });
